Cleaned up & broke components & functions into multiple files
